Move "scopes" tests in their right place
