+++ ban/tests/http/test_auth.py
@@ -1,5 +1,6 @@
 import pytest
 
+from ban.auth import models as amodels
 from ..factories import ClientFactory, UserFactory
 
 
@@ -82,3 +83,29 @@ def test_can_request_token_with_json_enoded_body(client):
     }, content_type='application/json')
     assert resp.status_code == 200
     assert 'access_token' in resp.json
+
+
+def test_create_token_with_scopes(client):
+    c = ClientFactory(scopes=['municipality_write'])
+    resp = client.post('/token/', data={
+        'grant_type': 'client_credentials',
+        'client_id': str(c.client_id),
+        'client_secret': c.client_secret,
+        'ip': '1.2.3.4',
+    })
+    assert resp.status_code == 200
+    token = amodels.Token.first()
+    assert token.scopes == ['municipality_write']
+
+
+def test_create_token_without_scopes(client):
+    c = ClientFactory(scopes=None)
+    resp = client.post('/token/', data={
+        'grant_type': 'client_credentials',
+        'client_id': str(c.client_id),
+        'client_secret': c.client_secret,
+        'ip': '1.2.3.4',
+    })
+    assert resp.status_code == 200
+    token = amodels.Token.first()
+    assert token.scopes == []
