+++ ban/tests/http/test_municipality.py
@@ -453,3 +453,121 @@ def test_authorized_responses_contain_sessions_data(get):
     session = context.get('session')
     assert resp.headers['Session-Client'] == session.client.id
     assert resp.headers['Session-User'] == session.user.id
+
+
+@authorize('view')
+def test_get_municipality_without_scopes(get):
+    municipality = MunicipalityFactory(name="Cabour")
+    resp = get('/municipality/{}'.format(municipality.id))
+    assert resp.status_code == 200
+    assert resp.json['id']
+    assert resp.json['name'] == 'Cabour'
+
+
+@authorize('view')
+def test_post_municipality_without_scopes(post):
+    data = {
+        "name": "Fornex",
+        "insee": "12345",
+        "siren": '123456789',
+    }
+    resp = post('/municipality', data)
+    assert resp.status_code == 401
+
+
+@authorize('municipality_write')
+def test_post_municipality_with_scopes(post):
+    data = {
+        "name": "Fornex",
+        "insee": "12345",
+        "siren": '123456789',
+    }
+    resp = post('/municipality', data)
+    assert resp.status_code == 201
+    assert resp.json['id']
+    assert resp.json['name'] == 'Fornex'
+    assert models.Municipality.select().count() == 1
+
+
+@authorize('postcode_write')
+def test_cannot_post_municipality_with_postcode_scope(post):
+    data = {
+        "name": "Fornex",
+        "insee": "12345",
+        "siren": '123456789',
+    }
+    resp = post('/municipality', data)
+    assert resp.status_code == 401
+    assert models.Municipality.select().count() == 0
+
+
+@authorize('whatever_but_municipality_write')
+def test_cannot_patch_municipality_without_scopes(patch):
+    municipality = MunicipalityFactory()
+    data = {
+        "version": 2,
+        "alias": ["Martigues"]
+    }
+    resp = patch('/municipality/{}'.format(municipality.id), data)
+    assert resp.status_code == 401
+    municipality = models.Municipality.first()
+    assert municipality.alias == []
+
+
+@authorize('municipality_write')
+def test_patch_municipality_with_scopes(patch):
+    municipality = MunicipalityFactory()
+    data = {
+        "version": 2,
+        "alias": ["Martigues"]
+    }
+    resp = patch('/municipality/{}'.format(municipality.id), data)
+    assert resp.status_code == 200
+    municipality = models.Municipality.first()
+    assert 'Martigues' in municipality.alias
+
+
+@authorize('view')
+def test_cannot_put_municipality_without_scopes(put):
+    municipality = MunicipalityFactory(name="SuperVille")
+    data = {
+        "name": "Martigues",
+        "insee": "13500",
+        "siren": '123456789',
+        "version": 2
+    }
+    resp = put('/municipality/{}'.format(municipality.id), data)
+    assert resp.status_code == 401
+    municipality = models.Municipality.first()
+    assert 'SuperVille' == municipality.name
+
+
+@authorize('municipality_write')
+def test_put_municipality_with_scopes(put):
+    municipality = MunicipalityFactory(name="SuperVille")
+    data = {
+        "name": "Martigues",
+        "insee": "13500",
+        "siren": '123456789',
+        "version": 2
+    }
+    resp = put('/municipality/{}'.format(municipality.id), data)
+    assert resp.status_code == 200
+    municipality = models.Municipality.first()
+    assert 'Martigues' == municipality.name
+
+
+@authorize('view')
+def test_cannot_delete_municipality_without_scopes(client):
+    municipality = MunicipalityFactory()
+    resp = client.delete('/municipality/{}'.format(municipality.id))
+    assert resp.status_code == 401
+    assert models.Municipality.select().count() == 1
+
+
+@authorize('municipality_write')
+def test_delete_municipality_with_scopes(client):
+    municipality = MunicipalityFactory()
+    resp = client.delete('/municipality/{}'.format(municipality.id))
+    assert resp.status_code == 200
+    assert not models.Municipality.select().count()
